# docs/event-service-openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
...
    "title" : "Event Service API",
    "version" : "1.0"
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/events/summary" : {
      "get" : {
        "description" : "Roles: ADMIN, ADMINISTRATIVE_STAFF.",
        "operationId" : "overview",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventsOverview"
                }
              }
            },
            "description" : "Totals by event status and registration status"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UNAUTHORIZED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "FORBIDDEN: caller is not ADMIN or ADMINISTRATIVE_STAFF"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INTERNAL_ERROR: unexpected error, no internal details are returned"
          }
        },
        "summary" : "Overall event and registration totals",
        "tags" : [ "Summaries" ]
      }
    },
    "/api/events/{eventId}/registrations" : {
      "get" : {
        "description" : "Roles: the owning EVENT_ORGANIZER or ACADEMIC_STAFF, or ADMIN / ADMINISTRATIVE_STAFF.",
        "operationId" : "eventSummary",
        "parameters" : [ {
          "in" : "path",
          "name" : "eventId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventRegistrationSummary"
                }
              }
            },
            "description" : "Confirmed and cancelled counts and remaining seats"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INVALID_PARAMETER: eventId is not a UUID"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UNAUTHORIZED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "FORBIDDEN: caller does not own the event"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "NOT_FOUND"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INTERNAL_ERROR: unexpected error, no internal details are returned"
          }
        },
        "summary" : "Registration and capacity summary for one event",
        "tags" : [ "Summaries" ]
      },
      "post" : {
        "description" : "Any signed-in user. Checks, in order: event is PUBLISHED, registration window is open, Group 5 eligibility (skipped for {\"all\": true} events), capacity.",
        "operationId" : "register",
        "parameters" : [ {
          "in" : "path",
          "name" : "eventId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RegistrationResponse"
                }
              }
            },
            "description" : "Registration CONFIRMED"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "EVENT_NOT_PUBLISHED, REGISTRATION_CLOSED or INVALID_PARAMETER (eventId is not a UUID)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UNAUTHORIZED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "NOT_ELIGIBLE: Group 5 says the caller does not meet the event's eligibility rule (message explains why), or INVALID_USER: Group 5 does not know the caller or the account is inactive"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "NOT_FOUND: event does not exist"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "CAPACITY_REACHED or ALREADY_REGISTERED"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INTERNAL_ERROR: unexpected error, no internal details are returned"
          },
          "503" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "GROUP5_UNAVAILABLE: eligibility service unreachable, nothing was saved"
          }
        },
        "summary" : "Register the caller for an event",
        "tags" : [ "Registrations" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ApiErrorResponse" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/ErrorDetail"
          },
          "success" : {
            "type" : "boolean",
            "example" : false
          }
        }
      },
...
      "ErrorDetail" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "example" : "CAPACITY_REACHED"
          },
          "message" : {
            "type" : "string",
            "example" : "This event has reached its registration capacity."
          }
        }
      },
      "EventRegistrationSummary" : {
        "type" : "object",
        "properties" : {
          "cancelled" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 3
          },
          "capacity" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 30
          },
          "confirmed" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Registrations that hold a seat",
            "example" : 12
          },
          "eventId" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "a0000000-0000-0000-0000-000000000001"
          },
          "remainingSeats" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "capacity - confirmed, never below 0",
            "example" : 18
          },
          "status" : {
            "type" : "string",
            "enum" : [ "DRAFT", "PUBLISHED", "CANCELLED", "COMPLETED" ]
          },
          "title" : {
            "type" : "string",
            "example" : "Innovation Week Workshop"
          }
        }
      },
...
      "EventsOverview" : {
        "type" : "object",
        "properties" : {
          "cancelled" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 1
          },
          "cancelledRegistrations" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 5
          },
          "completed" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 2
          },
          "confirmedRegistrations" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 40
          },
          "draft" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 1
          },
          "published" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 3
          },
          "totalEvents" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "draft + published + cancelled + completed",
            "example" : 7
          }
        }
      },
      "RegistrationResponse" : {
        "type" : "object",
        "properties" : {
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "eventId" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "a0000000-0000-0000-0000-000000000001"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "b0000000-0000-0000-0000-000000000001"
          },
          "status" : {
            "type" : "string",
            "description" : "CONFIRMED counts toward capacity; CANCELLED by the user; WAITLISTED is reserved and not used yet",
            "enum" : [ "CONFIRMED", "CANCELLED", "WAITLISTED" ]
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "userId" : {
            "type" : "string",
            "description" : "Group 5 user id of the registered user",
            "example" : "usr-student-001"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}